# swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Balance Service",
    "version": "1.0.11",
    "description": "This is a service which work with user's and organization's balances. You can to add money on your personal account or your organization. Also you can transfer money between accounts. You can buy the services from other organizations."
  },
  "servers": [
    {
      "url": "http://localhost:8081/"
    }
  ],
  "tags": [
    {
      "name": "user",
      "description": "Operations about user"
    },
    {
      "name": "operations",
      "description": "Available operations"
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get users list",
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Users"
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          }
        }
      }
    },
    "/user/{uuid}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get user by user id",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "description": "Uuid is an unique value",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          }
        }
      }
    },
    "/user": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Create user",
        "operationId": "createUser",
        "requestBody": {
          "description": "Create a new user in service",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          }
        }
      }
    },
    "/addition": {
      "patch": {
        "tags": [
          "operations"
        ],
        "summary": "Add money on the user's balance",
        "operationId": "addMoney",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credition"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Complete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Messages"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          }
        }
      }
    },
    "/transfer": {
      "patch": {
        "tags": [
          "operations"
        ],
        "summary": "Transfer money from users",
        "operationId": "transferMoney",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Transfer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Complete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Messages"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          }
        }
      }
    },
    "/buy": {
      "patch": {
        "tags": [
          "operations"
        ],
        "summary": "Buy services",
        "operationId": "byServices",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Buyer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Complete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Messages"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Users": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "uid": {
              "type": "string"
            },
            "name": {
              "type": "string"
            }
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "rub": {
            "type": "integer"
          },
          "pen": {
            "type": "integer"
          },
          "res_rub": {
            "type": "integer"
          },
          "res_pen": {
            "type": "integer"
          }
        }
      },
      "Credition": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string"
          },
          "rub": {
            "type": "integer"
          },
          "pen": {
            "type": "integer"
          }
        }
      },
      "Transfer": {
        "type": "object",
        "properties": {
          "uuid_sender": {
            "type": "string"
          },
          "uuid_reciever": {
            "type": "string"
          },
          "rub": {
            "type": "integer"
          },
          "pen": {
            "type": "integer"
          }
        }
      },
      "Buyer": {
        "type": "object",
        "properties": {
          "uuid_buyer": {
            "type": "string"
          },
          "uuid_services": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Messages": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "Errors": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  }
}